+++ src/lib/primevue/components/checkbox/colors/light/index.ts
@@ -1,30 +1,30 @@
 const lightColors = {
   root: {
-    background: '{form.field.background}',
-    checkedBackground: '{primary.color}',
-    checkedHoverBackground: '{primary.hover.color}',
-    disabledBackground: '{form.field.disabled.background}',
+    background: '{transparent}',
+    checkedBackground: '{surface.700}',
+    checkedHoverBackground: '{surface.600}',
+    disabledBackground: '{surface.200}',
     filledBackground: '{form.field.filled.background}',
-    borderColor: '{form.field.border.color}',
-    hoverBorderColor: '{form.field.hover.border.color}',
-    focusBorderColor: '{form.field.border.color}',
-    checkedBorderColor: '{primary.color}',
-    checkedHoverBorderColor: '{primary.hover.color}',
-    checkedFocusBorderColor: '{primary.color}',
-    checkedDisabledBorderColor: '{form.field.border.color}',
+    borderColor: '{surface.700}',
+    hoverBorderColor: '{surface.600}',
+    focusBorderColor: '{focus-color}',
+    checkedBorderColor: '{surface.700}',
+    checkedHoverBorderColor: '{surface.600}',
+    checkedFocusBorderColor: '{focus-color}',
+    checkedDisabledBorderColor: '{surface.200}',
     invalidBorderColor: '{form.field.invalid.border.color}',
     shadow: '{form.field.shadow}',
     focusRing: {
-        color: '{focus.ring.color}',
-        shadow: '{focus.ring.shadow}'
+      color: '{focus-color}',
+      shadow: '{focus.ring.shadow}'
     },
     transitionDuration: '{form.field.transition.duration}',
   },
   icon: {
     color: '{form.field.color}',
-    checkedColor: '{primary.contrast.color}',
-    checkedHoverColor: '{primary.contrast.color}',
-    disabledColor: '{form.field.disabled.color}',
+    checkedColor: '{surface.0}',
+    checkedHoverColor: '{surface.0}',
+    disabledColor: '{surface.400}',
   }
 }