+++ src/lib/primevue/components/checkbox/sizes/index.ts
@@ -1,15 +1,17 @@
 const sizes = {
   root: {
     borderRadius: '{border.radius.sm}',
-    width: '1.25rem',
-    height: '1.25rem',
+    width: '1.125rem',
+    height: '1.125rem',
+    wrapperWidth: '1.5rem',
+    wrapperHeight: '1.5rem',
     sm: {
       width: '1rem',
       height: '1rem'
     },
     lg: {
-      width: '1.5rem',
-      height: '1.5rem'
+      width: '1.10rem',
+      height: '1.10rem'
     },
     focusRing: {
       width: '{focus.ring.width}',
@@ -18,7 +20,7 @@ const sizes = {
     }
   },
   icon: {
-    size: '0.875rem',
+    size: '0.75rem',
     sm: {
       size: '0.75rem'
     },