+++ src/lib/primevue/components/index.ts
@@ -7,6 +7,11 @@ import ToolbarScheme from './toolbar';
 import ChipScheme from "./chip";
 import CheckboxScheme from "./checkbox";
 import TableScheme from "./table";
+import SwitcherScheme from "./switcher";
+import DividerScheme from "./divider";
+import SliderScheme from "./slider";
+import BreadcrumbScheme from "./breadcrumb";
+import RadioScheme from "./radio";
 
 export {
   AvatarScheme,
@@ -17,5 +22,10 @@ export {
   ToolbarScheme,
   ChipScheme,
   CheckboxScheme,
-  TableScheme
+  TableScheme,
+  SwitcherScheme,
+  DividerScheme,
+  SliderScheme,
+  BreadcrumbScheme,
+  RadioScheme,
 }