+++ src/lib/primevue/semantic/color-scheme/content/dark/index.ts
@@ -1,0 +1,10 @@
+export default {
+  content: {
+    background: '{surface.900}',
+    hoverBackground: '{surface.800}',
+    borderColor: '{surface.700}',
+    color: '{text.color}',
+    hoverColor: '{text.hoverColor}',
+    hoverMutedColor: '{text.hoverMutedColor}'
+  }
+}