+++ src/lib/primevue/semantic/color-scheme/content/light/index.ts
@@ -1,0 +1,10 @@
+export default {
+  content: {
+    background: '{surface.0}',
+    hoverBackground: '{surface.100}',
+    borderColor: '{surface.200}',
+    color: '{text.color}',
+    hoverColor: '{text.hoverColor}',
+    hoverMutedColor: '{text.hoverMutedColor}'
+  }
+}