+++ src/lib/primevue/semantic/color-scheme/form-field/dark/index.ts
@@ -1,0 +1,23 @@
+export default {
+  formField: {
+    background: '{surface.950}',
+    disabledBackground: '{surface.700}',
+    filledBackground: '{surface.800}',
+    filledHoverBackground: '{surface.800}',
+    filledFocusBackground: '{surface.800}',
+    borderColor: '{surface.600}',
+    hoverBorderColor: '{surface.500}',
+    focusBorderColor: '{primary.color}',
+    invalidBorderColor: '{error.color}',
+    color: '{surface.0}',
+    disabledColor: '{surface.400}',
+    placeholderColor: '{surface.400}',
+    invalidPlaceholderColor: '{red.400}',
+    floatLabelColor: '{surface.400}',
+    floatLabelFocusColor: '{primary.color}',
+    floatLabelActiveColor: '{surface.400}',
+    floatLabelInvalidColor: '{form.field.invalid.placeholder.color}',
+    iconColor: '{surface.400}',
+    shadow: '0 0 #0000, 0 0 #0000, 0 1px 2px 0 rgba(18, 18, 23, 0.05)'
+  }
+}